fix(app): move carousel constants to own file for fast-refresh
